+++ trunk/extensions/gtk_edit_element/gtk_edit_impl.h
@@ -78,8 +78,8 @@ class GtkEditImpl {
   Color GetTextColor();
   void SetFontFamily(const char *font);
   std::string GetFontFamily();
-  void SetFontSize(int size);
-  int GetFontSize();
+  void SetFontSize(double size);
+  double GetFontSize();
   void SetPasswordChar(const char *c);
   std::string GetPasswordChar();
   bool IsScrollBarRequired();
